+++ build.sh
@@ -1,7 +1,17 @@
 #!/bin/bash
-if [ ! -f packages/FAKE/tools/FAKE.exe ]; then
-  mono .nuget/NuGet.exe install FAKE -OutputDirectory packages -ExcludeVersion
+
+mono .paket/paket.bootstrapper.exe
+exit_code=$?
+if [ $exit_code -ne 0 ]; then
+	exit $exit_code
 fi
+
+mono .paket/paket.exe restore -v
+exit_code=$?
+if [ $exit_code -ne 0 ]; then
+	exit $exit_code
+fi
+  
 #workaround assembly resolution issues in build.fsx
 export FSHARPI=`which fsharpi`
 cat - > fsharpi <<"EOF"