+++ src/canopy/reporters.fs
@@ -6,34 +6,37 @@ open types
 
 type IReporter =
    abstract member testStart : string -> unit
-   abstract member pass : unit -> unit
-   abstract member fail : Exception -> string -> byte [] -> unit
-   abstract member todo : unit -> unit
-   abstract member skip : unit -> unit
+   abstract member pass : string -> unit
+   abstract member fail : Exception -> string -> byte [] -> string -> unit
+   abstract member todo : string -> unit
+   abstract member skip : string -> unit
    abstract member testEnd : string -> unit
    abstract member describe : string -> unit
    abstract member contextStart : string -> unit
    abstract member contextEnd : string -> unit
-   abstract member summary : int -> int -> int -> int -> unit
+   abstract member summary : int -> int -> int -> int -> int -> unit
    abstract member write : string -> unit
    abstract member suggestSelectors : string -> string list -> unit
    abstract member quit : unit -> unit
    abstract member suiteBegin : unit -> unit
    abstract member suiteEnd : unit -> unit
-   abstract member coverage : string -> byte [] -> unit
+   abstract member coverage : string -> byte [] -> string -> unit
+   abstract member setEnvironment : string -> unit
 
 type ConsoleReporter() =   
     interface IReporter with               
-        member this.pass () = 
+        member this.pass _ = 
             Console.ForegroundColor <- ConsoleColor.Green
             Console.WriteLine("Passed");
             Console.ResetColor()
 
-        member this.fail ex id ss = 
+        member this.fail ex id ss url = 
             Console.ForegroundColor <- ConsoleColor.Red
             Console.WriteLine("Error: ");
             Console.ResetColor()
             Console.WriteLine(ex.Message);
+            Console.Write("Url: ");
+            Console.WriteLine(url);
             Console.WriteLine("Stack: ");
             ex.StackTrace.Split([| "\r\n"; "\n" |], StringSplitOptions.None)
             |> Array.iter (fun trace -> 
@@ -58,12 +61,16 @@ type ConsoleReporter() =
         
         member this.contextEnd c = ()
         
-        member this.summary minutes seconds passed failed =
+        member this.summary minutes seconds passed failed skipped =
             Console.WriteLine()
             Console.WriteLine("{0} minutes {1} seconds to execute", minutes, seconds)
             if failed = 0 then
                 Console.ForegroundColor <- ConsoleColor.Green
             Console.WriteLine("{0} passed", passed)
+            Console.ResetColor()
+            if skipped > 0 then
+                Console.ForegroundColor <- ConsoleColor.Yellow        
+            Console.WriteLine("{0} skipped", skipped)    
             Console.ResetColor()
             if failed > 0 then
                 Console.ForegroundColor <- ConsoleColor.Red        
@@ -73,7 +80,7 @@ type ConsoleReporter() =
         member this.write w = Console.WriteLine w        
         
         member this.suggestSelectors selector suggestions = 
-            Console.ForegroundColor <- ConsoleColor.DarkYellow                    
+            Console.ForegroundColor <- ConsoleColor.Yellow                    
             Console.WriteLine("Couldn't find any elements with selector '{0}', did you mean:", selector)
             suggestions |> List.iter (fun suggestion -> Console.WriteLine("\t{0}", suggestion))
             Console.ResetColor()
@@ -90,11 +97,16 @@ type ConsoleReporter() =
 
         member this.suiteEnd () = ()
 
-        member this.coverage url ss = ()
+        member this.coverage url ss _ = ()
         
-        member this.todo () = ()
+        member this.todo _ = ()
 
-        member this.skip () = ()
+        member this.skip id =
+            Console.ForegroundColor <- ConsoleColor.Yellow
+            Console.WriteLine("Skipped");
+            Console.ResetColor()
+
+        member this.setEnvironment env = ()
 
 type TeamCityReporter() =
 
@@ -117,9 +129,9 @@ type TeamCityReporter() =
         consoleReporter.describe temcityReport
 
     interface IReporter with               
-        member this.pass () = consoleReporter.pass ()
+        member this.pass id = consoleReporter.pass id
     
-        member this.fail ex id ss =
+        member this.fail ex id ss url =
             let mutable image = ""
             if not (Array.isEmpty ss) then
                 image <- String.Format("canopy-image({0})", Convert.ToBase64String(ss))
@@ -128,7 +140,7 @@ type TeamCityReporter() =
                                 (tcFriendlyMessage id) 
                                 (tcFriendlyMessage ex.Message)    
                                 (tcFriendlyMessage image))
-            consoleReporter.fail ex id ss
+            consoleReporter.fail ex id ss url
     
         member this.describe d = 
             teamcityReport (sprintf "message text='%s' status='NORMAL'" (tcFriendlyMessage d))
@@ -142,7 +154,7 @@ type TeamCityReporter() =
             teamcityReport (sprintf "testSuiteFinished name='%s'" (tcFriendlyMessage c))
             consoleReporter.contextEnd c
     
-        member this.summary minutes seconds passed failed = consoleReporter.summary minutes seconds passed failed        
+        member this.summary minutes seconds passed failed skipped = consoleReporter.summary minutes seconds passed failed skipped        
     
         member this.write w = consoleReporter.write w        
     
@@ -158,11 +170,13 @@ type TeamCityReporter() =
 
         member this.suiteEnd () = ()
         
-        member this.coverage url ss = ()
+        member this.coverage url ss _ = ()
 
-        member this.todo () = ()
+        member this.todo _ = ()
 
-        member this.skip () = ()
+        member this.skip id = teamcityReport (sprintf "testIgnored name='%s'" (tcFriendlyMessage id))
+
+        member this.setEnvironment env = ()
 
 type LiveHtmlReporter(browser : BrowserStartMode, driverPath : string) =
     let consoleReporter : IReporter = new ConsoleReporter() :> IReporter
@@ -181,10 +195,13 @@ type LiveHtmlReporter(browser : BrowserStartMode, driverPath : string) =
         | ChromeWithOptionsAndTimeSpan(options, timeSpan) -> new OpenQA.Selenium.Chrome.ChromeDriver(driverPath, options, timeSpan) :> IWebDriver
         | Firefox -> new OpenQA.Selenium.Firefox.FirefoxDriver() :> IWebDriver
         | FirefoxWithProfile profile -> new OpenQA.Selenium.Firefox.FirefoxDriver(profile) :> IWebDriver
-        | FirefoxWithPath path -> new OpenQA.Selenium.Firefox.FirefoxDriver(Firefox.FirefoxBinary(path), Firefox.FirefoxProfile()) :> IWebDriver
+        | FirefoxWithPath path -> new OpenQA.Selenium.Firefox.FirefoxDriver(new Firefox.FirefoxBinary(path), Firefox.FirefoxProfile()) :> IWebDriver
         | ChromeWithUserAgent userAgent -> raise(System.Exception("Sorry ChromeWithUserAgent can't be used for LiveHtmlReporter"))
         | FirefoxWithUserAgent userAgent -> raise(System.Exception("Sorry FirefoxWithUserAgent can't be used for LiveHtmlReporter"))
+        | FirefoxWithPathAndTimeSpan(path, timespan) -> new OpenQA.Selenium.Firefox.FirefoxDriver(new Firefox.FirefoxBinary(path), Firefox.FirefoxProfile(), timespan) :> IWebDriver
+        | Safari -> new OpenQA.Selenium.Safari.SafariDriver() :> IWebDriver
         | PhantomJS | PhantomJSProxyNone -> raise(System.Exception("Sorry PhantomJS can't be used for LiveHtmlReporter"))
+        | Remote(_,_) -> raise(System.Exception("Sorry Remote can't be used for LiveHtmlReporter"))
                 
     let pin () =   
         let h = System.Windows.Forms.Screen.PrimaryScreen.WorkingArea.Height;
@@ -194,10 +211,13 @@ type LiveHtmlReporter(browser : BrowserStartMode, driverPath : string) =
         _browser.Manage().Window.Position <- new System.Drawing.Point((maxWidth * 0),0);
     do pin()
        
-    let mutable context = System.String.Empty;
-    let mutable test = System.String.Empty;
+    let mutable context = System.String.Empty
     let mutable canQuit = false
     let mutable contexts : string list = []
+    let mutable environment = System.String.Empty
+    let testStopWatch = System.Diagnostics.Stopwatch()
+    let contextStopWatch = System.Diagnostics.Stopwatch()
+    let jsEncode value = System.Web.HttpUtility.JavaScriptStringEncode(value)
 
     new() = LiveHtmlReporter(Firefox, String.Empty)
     new(browser : BrowserStartMode) = LiveHtmlReporter(browser, String.Empty)
@@ -206,7 +226,7 @@ type LiveHtmlReporter(browser : BrowserStartMode, driverPath : string) =
         with get () = _browser
 
     member val reportPath = None with get, set
-    member val reportTemplateUrl = @"http://lefthandedgoat.github.com/canopy/reporttemplate.html" with get, set
+    member val reportTemplateUrl = @"http://lefthandedgoat.github.com/canopy/reporttemplatep.html" with get, set
     member this.js script = (_browser :?> IJavaScriptExecutor).ExecuteScript(script)
     member this.reportHtml () = (this.js "return $('*').html();").ToString()
     member private this.swallowedJS script = try (_browser :?> IJavaScriptExecutor).ExecuteScript(script) |> ignore with | ex -> ()
@@ -215,43 +235,106 @@ type LiveHtmlReporter(browser : BrowserStartMode, driverPath : string) =
             then System.IO.Directory.CreateDirectory(directory) |> ignore
         IO.File.WriteAllText(System.IO.Path.Combine(directory,filename + ".html"), this.reportHtml())
     
+    member this.commonFail ctx (ex:Exception) id ss url =
+        let encodedId = jsEncode id
+        this.swallowedJS (sprintf "updateTestInContext('%s', '%s','Fail', '%s');" ctx encodedId (Convert.ToBase64String(ss)))
+        let stack = sprintf "%s%s%s" ex.Message System.Environment.NewLine ex.StackTrace
+        let stack = jsEncode stack
+        this.swallowedJS (sprintf "addStackToTest ('%s', '%s', '%s');" ctx encodedId stack)
+        this.swallowedJS (sprintf "addUrlToTest ('%s', '%s', '%s');" ctx encodedId url)
+        consoleReporter.fail ex id ss url
+        
+    member this.passWithContext ctx id =
+        let encodedId = jsEncode id
+        let context = jsEncode ctx
+        this.swallowedJS (sprintf "updateTestInContext('%s', '%s', 'Pass', '%s');" context encodedId "")
+        consoleReporter.pass id
+
+    member this.failWithContext ctx ex id ss url = 
+        let context = jsEncode ctx
+        this.commonFail context ex id ss url
+
+    member this.writeWithContext ctx w id =
+        let encodedId = jsEncode id
+        let encoded = jsEncode w
+        let context = jsEncode ctx
+        this.swallowedJS (sprintf "addMessageToTestByName ('%s', '%s', '%s');" context encodedId encoded)
+        consoleReporter.write w
+
+    member this.testStartWithContext ctx id =
+        let encodedId = jsEncode id
+        let context = jsEncode ctx
+        this.swallowedJS (sprintf "addToContext ('%s', '%s');" context encodedId)
+        consoleReporter.testStart id
+            
+    member this.testEndWithContext ctx id minutes seconds =
+        let encodedId = jsEncode id
+        let context = jsEncode ctx
+        this.swallowedJS (sprintf "addTimeToTest ('%s', '%s', '%im %is');" context encodedId minutes seconds)
+
+    member this.todoWithContext ctx id =
+        let encodedId = jsEncode id
+        let context = jsEncode ctx
+        this.swallowedJS (sprintf "updateTestInContext('%s', '%s', 'Todo', '%s');" context encodedId "")
+
+    member this.skipWithContext ctx id =
+        let encodedId = jsEncode id
+        let context = jsEncode ctx
+        this.swallowedJS (sprintf "updateTestInContext('%s', '%s', 'Skip', '%s');" context encodedId "")
+        consoleReporter.skip id
 
     interface IReporter with               
-        member this.pass () =
-            this.swallowedJS (sprintf "addToContext('%s', 'Pass', '%s', '%s');" context test "")
-            consoleReporter.pass ()
+        member this.pass id =
+            let encodedId = jsEncode id
+            this.swallowedJS (sprintf "updateTestInContext('%s', '%s', 'Pass', '%s');" context encodedId "")
+            consoleReporter.pass id
 
-        member this.fail ex id ss =
-            this.swallowedJS (sprintf "addToContext('%s', 'Fail', '%s', '%s');" context test (Convert.ToBase64String(ss)))
-            consoleReporter.fail ex id ss
-
-        member this.describe d = 
+        member this.fail ex id ss url = this.commonFail context ex id ss url
+                
+        member this.describe d =
+            let encoded = jsEncode d
+            this.swallowedJS (sprintf "addMessageToTest ('%s', '%s');" context encoded) 
             consoleReporter.describe d
           
         member this.contextStart c = 
+            contextStopWatch.Reset()
+            contextStopWatch.Start()
             contexts <- c :: contexts
-            context <- System.Web.HttpUtility.JavaScriptStringEncode(c)
+            context <- jsEncode c
             this.swallowedJS (sprintf "addContext('%s');" context)
             this.swallowedJS (sprintf "collapseContextsExcept('%s');" context)
             consoleReporter.contextStart c
 
         member this.contextEnd c = 
+            contextStopWatch.Stop()
+            let ellapsed = contextStopWatch.Elapsed
+            this.swallowedJS (sprintf "addTimeToContext ('%s', '%im %is');" context ellapsed.Minutes ellapsed.Seconds)
             consoleReporter.contextEnd c
 
-        member this.summary minutes seconds passed failed =                        
-            consoleReporter.summary minutes seconds passed failed
+        member this.summary minutes seconds passed failed skipped =
+            this.swallowedJS (sprintf "setTotalTime ('%im %is');" minutes seconds)                        
+            consoleReporter.summary minutes seconds passed failed skipped
         
         member this.write w = 
+            let encoded = jsEncode w
+            this.swallowedJS (sprintf "addMessageToTest ('%s', '%s');" context encoded)
             consoleReporter.write w
         
         member this.suggestSelectors selector suggestions = 
             consoleReporter.suggestSelectors selector suggestions
 
         member this.testStart id = 
-            test <- System.Web.HttpUtility.JavaScriptStringEncode(id)
+            testStopWatch.Reset()
+            testStopWatch.Start()
+            let encodedId = jsEncode id
+            this.swallowedJS (sprintf "addToContext ('%s', '%s');" context encodedId)
             consoleReporter.testStart id
             
-        member this.testEnd id = ()
+        member this.testEnd id =
+            let encodedId = jsEncode id
+            testStopWatch.Stop()
+            let ellapsed = testStopWatch.Elapsed
+            this.swallowedJS (sprintf "addTimeToTest ('%s', '%s', '%im %is');" context encodedId ellapsed.Minutes ellapsed.Seconds)
 
         member this.quit () = 
           match this.reportPath with
@@ -262,20 +345,31 @@ type LiveHtmlReporter(browser : BrowserStartMode, driverPath : string) =
 
           if canQuit then _browser.Quit()
         
-        member this.suiteBegin () = _browser.Navigate().GoToUrl(this.reportTemplateUrl)
+        member this.suiteBegin () = 
+            _browser.Navigate().GoToUrl(this.reportTemplateUrl)
+            this.swallowedJS (sprintf "setStartTime ('%s');" (String.Format("{0:F}", System.DateTime.Now)))    
+            if environment <> String.Empty then this.swallowedJS (sprintf "setEnvironment ('%s');" environment)             
 
         member this.suiteEnd () = 
             canQuit <- true
             this.swallowedJS (sprintf "collapseContextsExcept('%s');" "") //cheap hack to collapse all contexts at the end of a run
 
-        member this.coverage url ss = 
+        member this.coverage url ss id =
+            let encodedId = jsEncode id
             if (contexts |> List.exists (fun c -> c = "Coverage Reports")) = false then
                 contexts <- "Coverage Reports" :: contexts
                 this.swallowedJS (sprintf "addContext('%s');" "Coverage Reports")
-            this.swallowedJS (sprintf "addToContext('%s', 'Pass', '%s', '%s');" "Coverage Reports" url (Convert.ToBase64String(ss)))
+            this.swallowedJS (sprintf "addToContext ('%s', '%s');" "Coverage Reports" url)
+            this.swallowedJS (sprintf "updateTestInContext('%s', '%s', 'Pass', '%s');" "Coverage Reports" encodedId (Convert.ToBase64String(ss)))
 
-        member this.todo () = 
-            this.swallowedJS (sprintf "addToContext('%s', 'Todo', '%s', '%s');" context test "")
+        member this.todo id =
+            let encodedId = jsEncode id
+            this.swallowedJS (sprintf "updateTestInContext('%s', '%s', 'Todo', '%s');" context encodedId "")
 
-        member this.skip () = 
-            this.swallowedJS (sprintf "addToContext('%s', 'Skip', '%s', '%s');" context test "")
+        member this.skip id =
+            let encodedId = jsEncode id
+            this.swallowedJS (sprintf "updateTestInContext('%s', '%s', 'Skip', '%s');" context encodedId "")
+            consoleReporter.skip id
+
+        member this.setEnvironment env =
+            environment <- env