+++ src/canopy/runner.fs
@@ -1,4 +1,5 @@
-﻿module canopy.runner
+﻿[<AutoOpen>]
+module canopy.runner
 
 open System
 open configuration
@@ -7,10 +8,9 @@ open reporters
 open types
 open OpenQA.Selenium
 
-let rec private last = function
-    | hd :: [] -> hd
-    | hd :: tl -> last tl
-    | _ -> failwith "Empty list."
+let private last = function
+    | hd :: tl -> hd
+    | [] -> failwith "Empty list."
 
 let mutable suites = [new suite()]
 let mutable todo = fun () -> ()
@@ -26,74 +26,136 @@ let context c =
     else 
         let s = new suite()
         s.Context <- c
-        suites <- suites @ [s]
-
-let ( &&& ) description f = (last suites).Tests <- (last suites).Tests @ [Test(description, f, (last suites).Tests.Length + 1)]
+        suites <- s::suites
+let private incrementLastTestSuite () =
+    let lastSuite = last suites
+    lastSuite.TotalTestsCount <- lastSuite.TotalTestsCount + 1
+    lastSuite
+let ( &&& ) description f = 
+    let lastSuite = incrementLastTestSuite()
+    lastSuite.Tests <- Test(description, f, lastSuite.TotalTestsCount)::lastSuite.Tests
 let test f = null &&& f
 let ntest description f = description &&& f
-let ( &&&& ) description f = (last suites).Wips <- (last suites).Wips @ [Test(description, f, (last suites).Wips.Length + 1)]
+let ( &&&& ) description f = 
+    let lastSuite = incrementLastTestSuite()
+    lastSuite.Wips <- Test(description, f, lastSuite.TotalTestsCount)::lastSuite.Wips
 let wip f = null &&&& f
-let many count f = [1 .. count] |> List.iter (fun _ -> (last suites).Manys <- (last suites).Manys @ [Test(null, f, (last suites).Manys.Length + 1)])
-let ( &&! ) description f = (last suites).Tests <- (last suites).Tests @ [Test(description, skipped, (last suites).Tests.Length + 1)]
+let many count f = 
+    let lastSuite = incrementLastTestSuite()
+    [1 .. count] |> List.iter (fun _ -> lastSuite.Manys <- Test(null, f, lastSuite.TotalTestsCount)::lastSuite.Manys)
+let nmany count description f = 
+    let lastSuite = incrementLastTestSuite()
+    [1 .. count] |> List.iter (fun _ -> lastSuite.Manys <- Test(description, f, lastSuite.TotalTestsCount)::lastSuite.Manys)
+let ( &&! ) description f = 
+    let lastSuite = incrementLastTestSuite()
+    lastSuite.Tests <- Test(description, skipped, lastSuite.TotalTestsCount)::lastSuite.Tests
 let xtest f = null &&! f
-
+let ( &&&&& ) description f = 
+    let lastSuite = incrementLastTestSuite()
+    lastSuite.Always <- Test(description, f, lastSuite.TotalTestsCount)::lastSuite.Always
 let mutable passedCount = 0
+let mutable skippedCount = 0
 let mutable failedCount = 0
 let mutable contextFailed = false
 let mutable failedContexts : string list = []
 let mutable failed = false
 
-let pass () =    
+let pass id =    
     passedCount <- passedCount + 1
-    reporter.pass ()
+    reporter.pass id
 
-let fail (ex : Exception) id =
-    try
-        if failFast = ref true then failed <- true        
-        failedCount <- failedCount + 1
-        contextFailed <- true
-        let f = DateTime.Now.ToString("MMM-d_HH-mm-ss-fff")
-        let ss = screenshot configuration.failScreenshotPath f
-        reporter.fail ex id ss
-    with 
-        | :? WebDriverException as failExc -> 
-            //Fail during error report (likely  OpenQA.Selenium.WebDriverException.WebDriverTimeoutException ). 
-            // Don't fail the runner itself, but report it.
-            reporter.write (sprintf "Error during fail reporting: %s" (failExc.ToString()))
-            reporter.fail ex id Array.empty
+let skip id =    
+    skippedCount <- skippedCount + 1
+    reporter.skip id
+
+let fail (ex : Exception) (test : Test) (suite : suite) autoFail url =
+    if failureMessagesThatShoulBeTreatedAsSkip |> List.exists (fun message -> ex.Message = message) then
+        skip test.Id
+    else
+        if skipAllTestsOnFailure = true || skipRemainingTestsInContextOnFailure = true then skipNextTest <- true
+        if autoFail then
+            skip test.Id //dont take the time to fail all the tests just skip them
+        else
+            try
+                if failFast = ref true then failed <- true        
+                failedCount <- failedCount + 1
+                contextFailed <- true
+                let f = configuration.failScreenshotFileName test suite
+                if failureScreenshotsEnabled = true then
+                  let ss = screenshot configuration.failScreenshotPath f
+                  reporter.fail ex test.Id ss url
+                else reporter.fail ex test.Id Array.empty<byte> url
+            with 
+                | failExc -> 
+                    //Fail during error report (likely  OpenQA.Selenium.WebDriverException.WebDriverTimeoutException ). 
+                    // Don't fail the runner itself, but report it.
+                    reporter.write (sprintf "Error during fail reporting: %s" (failExc.ToString()))
+                    reporter.fail ex test.Id Array.empty url
+
+let safelyGetUrl () = if browser = null then "no browser = no url" else browser.Url
 
 let failSuite (ex: Exception) (suite : suite) =    
     let reportFailedTest (ex: Exception) (test : Test) =
         reporter.testStart test.Id  
-        fail ex test.Id
+        fail ex test suite true <| safelyGetUrl()
+        reporter.testEnd test.Id
+
+    // tests are in reverse order and have to be reversed first
+    do suite.Tests
+    |> List.rev
+    |> List.iter (fun test -> reportFailedTest ex test)
+
+let tryTest test suite func =
+    try
+        func ()                  
+        Pass
+    with
+        | ex when failureMessage <> null && failureMessage = ex.Message -> Pass
+        | ex -> Fail ex
+
+let private processRunResult suite (test : Test) result = 
+    match result with
+    | Pass -> pass test.Id
+    | Fail ex -> fail ex test suite false <| safelyGetUrl()
+    | Skip -> skip test.Id
+    | Todo -> reporter.todo test.Id
+    | FailFast -> ()
+    | Failed -> ()
+
+let private runtest (suite : suite) (test : Test) =
+    if failed = false then             
+        reporter.testStart test.Id
+        let result = 
+          if System.Object.ReferenceEquals(test.Func, todo) then Todo
+          else if System.Object.ReferenceEquals(test.Func, skipped) then Skip
+          else if skipNextTest = true then Skip
+          else
+              let testResult =
+                let testResult = tryTest test suite (suite.Before >> test.Func)
+                match testResult with
+                | Fail(_) -> processRunResult suite test testResult; Failed
+                | _ -> testResult
+              
+              let afterResult = tryTest test suite (suite.After)
+              match testResult with
+              | Failed -> testResult
+              | _ -> afterResult
+
         reporter.testEnd test.Id 
-    suite.Tests |> List.iter (fun test -> reportFailedTest ex test)
+        failureMessage <- null
+        result
+    else
+        failureMessage <- null
+        FailFast
 
 let run () =
     reporter.suiteBegin()
     let stopWatch = new Diagnostics.Stopwatch()
     stopWatch.Start()      
-    
-    let runtest (suite : suite) (test : Test) =
-        if failed = false then            
-            try 
-                reporter.testStart test.Id  
-                if System.Object.ReferenceEquals(test.Func, todo) then 
-                    reporter.todo ()
-                else if System.Object.ReferenceEquals(test.Func, skipped) then 
-                    reporter.skip ()
-                else
-                    suite.Before ()
-                    test.Func ()
-                    suite.After ()
-                    pass()
-            with
-                | ex when failureMessage <> null && failureMessage = ex.Message -> pass()
-                | ex -> fail ex test.Id 
-            reporter.testEnd test.Id 
         
-        failureMessage <- null            
-
+    // suites list is in reverse order and have to be reversed before running the tests
+    suites <- List.rev suites
+        
     //run all the suites
     if runFailedContextsFirst = true then
         let failedContexts = history.get()
@@ -103,39 +165,44 @@ let run () =
 
     //run only wips if there are any
     if suites |> List.exists (fun s -> s.Wips.IsEmpty = false) then
-        suites <- suites |> List.filter (fun s -> s.Wips.IsEmpty = false)
+        suites <- suites |> List.filter (fun s -> s.Wips.IsEmpty = false || s.Always.IsEmpty = false)
 
     suites
     |> List.iter (fun s ->
+        if skipRemainingTestsInContextOnFailure = true && skipAllTestsOnFailure = false then skipNextTest <- false
         if failed = false then
             contextFailed <- false
             if s.Context <> null then reporter.contextStart s.Context
             try
                 s.Once ()
+            with 
+                | ex -> failSuite ex s
+            if failed = false then
                 if s.Wips.IsEmpty = false then
                     wipTest <- true
-                    s.Wips |> List.iter (fun w -> runtest s w)
+                    let tests = s.Wips @ s.Always |> List.sortBy (fun t -> t.Number)
+                    tests |> List.iter (fun w -> runtest s w |> processRunResult s w)
                     wipTest <- false
                 else if s.Manys.IsEmpty = false then
-                    s.Manys |> List.iter (fun m -> runtest s m)
+                    s.Manys |> List.rev |> List.iter (fun m -> runtest s m |> processRunResult s m)
                 else
-                    s.Tests |> List.iter (fun t -> runtest s t)
-                s.Lastly ()        
-            with 
-                | ex -> failSuite ex s                    
+                    let tests = s.Tests @ s.Always |> List.sortBy (fun t -> t.Number)
+                    tests |> List.iter (fun t -> runtest s t |> processRunResult s t)
+            s.Lastly ()                  
 
-            if contextFailed = true then failedContexts <- failedContexts @ [s.Context]
+            if contextFailed = true then failedContexts <- s.Context::failedContexts
             if s.Context <> null then reporter.contextEnd s.Context
     )
     
     history.save failedContexts
 
     stopWatch.Stop()    
-    reporter.summary stopWatch.Elapsed.Minutes stopWatch.Elapsed.Seconds passedCount failedCount 
+    reporter.summary stopWatch.Elapsed.Minutes stopWatch.Elapsed.Seconds passedCount failedCount skippedCount 
     reporter.suiteEnd()
 
 let runFor browsers =
-    let currentSuites = suites
+    // suites are in reverse order and have to be reversed before running the tests
+    let currentSuites = suites |> List.rev
     suites <- [new suite()]
     match box browsers with
         | :? (types.BrowserStartMode list) as browsers -> 
@@ -155,4 +222,4 @@ let runFor browsers =
                 once (fun _ -> switchTo browser)
                 suites <- suites @ currentSuites)
         | _ -> raise <| Exception "I dont know what you have given me"
-    run ()
+